+++ enterprise/cmd/frontend/internal/authz/bitbucketserver/provider_test.go
@@ -459,7 +459,7 @@ func newClient(t *testing.T, name string) (*bitbucketserver.Client, func()) {
 }
 
 func newProvider(cli *bitbucketserver.Client, db *sql.DB, ttl time.Duration) *Provider {
-	p := NewProvider(cli, db, ttl)
+	p := NewProvider(cli, db, ttl, DefaultHardTTL)
 	p.pageSize = 1       // Exercise pagination
 	p.store.block = true // Wait for first update to complete.
 	return p